recon_datavis/requirements.txt
h5py==3.1.0
matplotlib==3.3.4
numpy
pandas==1.1.5
pickleshare==0.7.5
Pillow==8.2.0
Pygments==2.9.0
PyYAML==5.4.1
requests==2.25.1
utm==0.7.0